requirements.txt
Flask>=2.2
streamlit>=1.20
sentence-transformers>=2.2
llama-cpp-python>=0.1.67
faiss-cpu>=1.7
pdfplumber>=0.7
python-pptx>=0.6
python-docx>=0.8
tiktoken>=0.4.0